# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VPS Management Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  uptime?: number;
  timestamp?: string;
}

export type HostCapabilitiesHostMode =
  (typeof HostCapabilitiesHostMode)[keyof typeof HostCapabilitiesHostMode];

export const HostCapabilitiesHostMode = {
  "bare-metal": "bare-metal",
  vm: "vm",
  container: "container",
  unknown: "unknown",
} as const;

export type HostCapabilitiesStartupMode =
  (typeof HostCapabilitiesStartupMode)[keyof typeof HostCapabilitiesStartupMode];

export const HostCapabilitiesStartupMode = {
  systemd: "systemd",
  init: "init",
  manual: "manual",
} as const;

export type HostCapabilitiesAccessMode =
  (typeof HostCapabilitiesAccessMode)[keyof typeof HostCapabilitiesAccessMode];

export const HostCapabilitiesAccessMode = {
  direct: "direct",
  nat: "nat",
  none: "none",
} as const;

export interface HostCapabilities {
  systemd: boolean;
  kvm: boolean;
  docker: boolean;
  publicIpv4: boolean;
  virtualization: boolean;
  mountPoints: string[];
  detectedAt: string;
  hostMode: HostCapabilitiesHostMode;
  startupMode: HostCapabilitiesStartupMode;
  accessMode: HostCapabilitiesAccessMode;
}

export interface HostInfo {
  hostname: string;
  os: string;
  kernel: string;
  arch: string;
  cpuCount: number;
  totalMemoryMb: number;
  totalDiskGb: number;
  publicIp?: string;
  uptime: number;
}

export interface SystemMetrics {
  cpuPercent: number;
  memoryPercent: number;
  memoryUsedMb: number;
  memoryTotalMb: number;
  diskPercent: number;
  diskUsedGb: number;
  diskTotalGb: number;
  networkRxBps: number;
  networkTxBps: number;
  timestamp: string;
}

export interface MetricsDataPoint {
  cpuPercent: number;
  memoryPercent: number;
  networkRxBps: number;
  networkTxBps: number;
  timestamp: string;
}

export interface MetricsHistory {
  dataPoints: MetricsDataPoint[];
}

export type InstanceStatus =
  (typeof InstanceStatus)[keyof typeof InstanceStatus];

export const InstanceStatus = {
  running: "running",
  stopped: "stopped",
  starting: "starting",
  stopping: "stopping",
  error: "error",
} as const;

export type InstanceType = (typeof InstanceType)[keyof typeof InstanceType];

export const InstanceType = {
  kvm: "kvm",
  docker: "docker",
  lxc: "lxc",
} as const;

export interface Instance {
  id: string;
  name: string;
  status: InstanceStatus;
  cpuCores: number;
  memoryMb: number;
  diskGb: number;
  ipAddress?: string;
  os: string;
  createdAt: string;
  startedAt?: string;
  sshPort?: number;
  type: InstanceType;
  tags?: string[];
}

export interface InstanceList {
  instances: Instance[];
  total: number;
}

export type CreateInstanceRequestType =
  (typeof CreateInstanceRequestType)[keyof typeof CreateInstanceRequestType];

export const CreateInstanceRequestType = {
  kvm: "kvm",
  docker: "docker",
  lxc: "lxc",
} as const;

export interface CreateInstanceRequest {
  name: string;
  /**
   * @minimum 1
   * @maximum 32
   */
  cpuCores: number;
  /** @minimum 128 */
  memoryMb: number;
  /** @minimum 1 */
  diskGb: number;
  os: string;
  type: CreateInstanceRequestType;
  tags?: string[];
}

export type LogEntryLevel = (typeof LogEntryLevel)[keyof typeof LogEntryLevel];

export const LogEntryLevel = {
  debug: "debug",
  info: "info",
  warn: "warn",
  error: "error",
} as const;

export interface LogEntry {
  id: string;
  timestamp: string;
  level: LogEntryLevel;
  message: string;
  instanceId?: string;
  source: string;
}

export interface LogList {
  logs: LogEntry[];
  total: number;
}

export interface InstallerScript {
  script: string;
  version: string;
  checksum?: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetMetricsHistoryParams = {
  limit?: number;
};

export type GetLogsParams = {
  limit?: number;
  level?: GetLogsLevel;
  instanceId?: string;
};

export type GetLogsLevel = (typeof GetLogsLevel)[keyof typeof GetLogsLevel];

export const GetLogsLevel = {
  debug: "debug",
  info: "info",
  warn: "warn",
  error: "error",
} as const;