# lib/api-zod/src/generated/types/hostCapabilities.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VPS Management Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { HostCapabilitiesAccessMode } from "./hostCapabilitiesAccessMode";
import type { HostCapabilitiesHostMode } from "./hostCapabilitiesHostMode";
import type { HostCapabilitiesStartupMode } from "./hostCapabilitiesStartupMode";

export interface HostCapabilities {
  systemd: boolean;
  kvm: boolean;
  docker: boolean;
  publicIpv4: boolean;
  virtualization: boolean;
  mountPoints: string[];
  detectedAt: string;
  hostMode: HostCapabilitiesHostMode;
  startupMode: HostCapabilitiesStartupMode;
  accessMode: HostCapabilitiesAccessMode;
}