# lib/api-zod/src/generated/types/hostInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VPS Management Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface HostInfo {
  hostname: string;
  os: string;
  kernel: string;
  arch: string;
  cpuCount: number;
  totalMemoryMb: number;
  totalDiskGb: number;
  publicIp?: string;
  uptime: number;
}