# lib/api-zod/src/generated/types/instance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VPS Management Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { InstanceStatus } from "./instanceStatus";
import type { InstanceType } from "./instanceType";

export interface Instance {
  id: string;
  name: string;
  status: InstanceStatus;
  cpuCores: number;
  memoryMb: number;
  diskGb: number;
  ipAddress?: string;
  os: string;
  createdAt: string;
  startedAt?: string;
  sshPort?: number;
  type: InstanceType;
  tags?: string[];
}