# lib/api-zod/src/generated/types/instanceList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VPS Management Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { Instance } from "./instance";

export interface InstanceList {
  instances: Instance[];
  total: number;
}