# lib/api-zod/src/generated/types/metricsDataPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VPS Management Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface MetricsDataPoint {
  cpuPercent: number;
  memoryPercent: number;
  networkRxBps: number;
  networkTxBps: number;
  timestamp: string;
}