# lib/api-zod/src/generated/types/systemMetrics.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VPS Management Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface SystemMetrics {
  cpuPercent: number;
  memoryPercent: number;
  memoryUsedMb: number;
  memoryTotalMb: number;
  diskPercent: number;
  diskUsedGb: number;
  diskTotalGb: number;
  networkRxBps: number;
  networkTxBps: number;
  timestamp: string;
}